% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipetitletoingredients.R
\name{recipe_title_to_ingredients}
\alias{recipe_title_to_ingredients}
\title{Predict ingredients for a given dish title}
\usage{
recipe_title_to_ingredients(input_titles, prediction_confidence = 0.5)
}
\arguments{
\item{input_titles}{a list of character strings. required}

\item{prediction_confidence}{decimal from 0 to 1.}
}
\value{
dataframe with given title, title recognized by model, and predicted ingredients
}
\description{
Predict ingredients for a given dish title
}
\examples{
recipe_title_to_ingredients(list("best ever carrot cake", "mom's pineapple zucchini bread"))
recipe_title_to_ingredients(input_titles = list("meatless stew","meatless beef stew","meatless meatlover's stew","meatless meatlovers beef stew"), prediction_confidence = 0.5)

recipe_title_to_ingredients(input_titles = list("meatless stew"), prediction_confidence = 0.7)
recipe_title_to_ingredients(input_titles = list("meatless stew"), prediction_confidence = 0.5)
recipe_title_to_ingredients(input_titles = list("meatless stew"), prediction_confidence = 0.3)

}
